factor out scriptable interpreter functiosn
