+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [1.3.0] - 2022-07-18
+## [1.3.0] - 2022-07-29
 ### Added
  - JShell interactive interpreter
+ - currentAPI interpreter variables with `FlatProgramAPI` instance
 
 
 ## [1.2.0] - 2022-06-11
