+++ src/main/java/rubydragon/GhidraInterpreter.java
@@ -18,16 +18,12 @@
 
 package rubydragon;
 
-import java.io.FileNotFoundException;
-import java.io.IOException;
 import java.io.InputStream;
 import java.io.PrintWriter;
 import java.util.List;
 
 import ghidra.app.plugin.core.console.CodeCompletion;
 import ghidra.app.plugin.core.interpreter.InterpreterConsole;
-import ghidra.app.script.GhidraScript;
-import ghidra.app.script.GhidraState;
 import ghidra.program.model.address.Address;
 import ghidra.program.model.listing.Program;
 import ghidra.program.util.ProgramLocation;
@@ -35,12 +31,14 @@ import ghidra.program.util.ProgramSelection;
 import ghidra.util.Disposable;
 
 /**
- * An interpreter that users can use interactively as well as to run custom
- * scripts.
+ * An interpreter that users can use interactively.
  *
  * This class provides a common base that all interpreters should implement in
  * order to fit in to the overall system. It provides a common wrapper around
  * the language-specific internals of a given language environment.
+ *
+ * Interpreters that also support scripts should inherit from
+ * ScriptableGhidraInterpreter instead.
  */
 public abstract class GhidraInterpreter implements Disposable {
 
@@ -50,22 +48,6 @@ public abstract class GhidraInterpreter implements Disposable {
 	public abstract void dispose();
 
 	/**
-	 * Loads a provided GhidraState into the interpreter.
-	 *
-	 * @param state The state to load.
-	 */
-	public void loadState(GhidraState state) {
-		updateHighlight(state.getCurrentHighlight());
-		updateLocation(state.getCurrentLocation());
-		updateSelection(state.getCurrentSelection());
-		updateProgram(state.getCurrentProgram());
-
-		// this has to happen after the location update
-		// since it clobbers the current address right now
-		updateAddress(state.getCurrentAddress());
-	}
-
-	/**
 	 * Get a list of completions for the given command prefix.
 	 *
 	 * @param cmd The command to try to complete.
@@ -73,31 +55,6 @@ public abstract class GhidraInterpreter implements Disposable {
 	 * @return A list of possible code completions.
 	 */
 	public abstract List<CodeCompletion> getCompletions(String cmd);
-
-	/**
-	 * Runs the given script with the arguments and state provided.
-	 *
-	 * The provided state is loaded into the interpreter at the beginning of
-	 * execution, and the values of the globals are then exported back into the
-	 * state after it completes.
-	 *
-	 * If the script cannot be found but the script is not running in headless mode,
-	 * the user will be prompted to ignore the error, which will cause the function
-	 * to simply continue instead of throwing an IllegalArgumentException.
-	 *
-	 * @throws IllegalArgumentException if the script does not exist
-	 * @throws IOException              if the script could not be read
-	 * @throws FileNotFoundException    if the script file wasn't found
-	 *
-	 * @param script          The script to run.
-	 *
-	 * @param scriptArguments The arguments to pass to the script.
-	 *
-	 * @param scriptState     The script to load before the script runs, and update
-	 *                        after the script finishes.
-	 */
-	public abstract void runScript(GhidraScript script, String[] scriptArguments, GhidraState scriptState)
-			throws IllegalArgumentException, FileNotFoundException, IOException;
 
 	/**
 	 * Sets the error output stream for this interpreter.
@@ -173,16 +130,4 @@ public abstract class GhidraInterpreter implements Disposable {
 	 * @param sel The new selection.
 	 */
 	public abstract void updateSelection(ProgramSelection sel);
-
-	/**
-	 * Updates a state with the current selection/location/etc. variables from the
-	 * interpreter.
-	 *
-	 * This is intended to be called after a call to runScript, to make sure that
-	 * any updates made to these variables during execution are reflected in the end
-	 * state.
-	 *
-	 * @param scriptState The state to update.
-	 */
-	public abstract void updateState(GhidraState scriptState);
 }
